requirements.txt
flask
pydantic
python-dotenv
requests
supabase
